% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rezrDF_easy.R, R/rezrObj_easy.R
\name{addField}
\alias{addField}
\alias{changeField}
\alias{addField.rezrDF}
\alias{changeField.rezrDF}
\alias{addField.rezrObj}
\alias{changeField.rezrObj}
\title{Shortcut functions for functions beginning with addField and changeField}
\usage{
addField(x, ..., foreign = F)

changeField(x, ..., foreign = F)

\method{addField}{rezrDF}(rezrDF, ..., foreign = F)

\method{changeField}{rezrDF}(rezrDF, ..., foreign = F)

\method{addField}{rezrObj}(rezrObj, ..., foreign = F)

\method{changeField}{rezrObj}(rezrObj, ..., foreign = F)
}
\arguments{
\item{...}{Arguments to be passed to functions such as those in see \link{addFieldLocal.rezrDF}, \link{addFieldForeign.rezrDF}, \link{addFieldLocal.rezrObj} and \link{addFieldForeign.rezrObj}.}

\item{foreign}{If TRUE, you are adding a foreign field. Normally set to false; you can just call \link{addFieldForeign.rezrDF} / \link{addFieldForeign.rezrObj} for foreign fields.}

\item{rezrDF}{The rezrDF to be modified.}
}
\value{
The rezrDF or rezrObj being modified.
}
\description{
Mostly a shortcut for \link{addFieldLocal} and \link{changeFieldLocal}.
}
