% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_easy.R
\name{addFieldForeign.rezrDF}
\alias{addFieldForeign.rezrDF}
\alias{changeFieldForeign.rezrDF}
\title{Easily add a field to / change a field in a rezrDF using information from another rezrDF}
\usage{
\method{addFieldForeign}{rezrDF}(
  targetDF,
  sourceDF,
  targetForeignKeyName,
  targetFieldName = "",
  sourceFieldName = "",
  type = "simple",
  fieldaccess = "flex",
  complexAction = NULL,
  targetNodeMap = NULL
)

\method{changeFieldForeign}{rezrDF}(
  targetDF,
  sourceDF,
  targetForeignKeyName,
  targetFieldName = "",
  sourceFieldName = "",
  type = "simple",
  fieldaccess = "flex",
  complexAction = NULL,
  targetNodeMap = NULL
)
}
\arguments{
\item{targetDF}{The rezrDF you will be changing.}

\item{sourceDF}{The rezrDF you will be getting information from.}

\item{targetForeignKeyName}{The name of the field in the target DF that corresponds to the key (ID) of the source DF. If type = "simple", this is a column in the rezrDF. If type = "complex", this is a field in the target node map, and targetNodeMap must be specified.}

\item{targetFieldName}{The name of the field you want to add / change.}

\item{sourceFieldName}{The name of the field in the source rezrDF that you will be pulling information from.}

\item{type}{If "simple", each row of the target DF gets information from only one row of the source DF. If "complex", each row of the target DF gets information from multiple rows of the source DF.}

\item{fieldaccess}{The field access value you want to set to the new / changed field.}

\item{complexAction}{When type = "complex", the action that will aggregate the values from the source DF together. See \link{complexActions} for ideas.}

\item{targetNodeMap}{The node map corresponding to the target DF. It only needs to be specified when type = "complex".}
}
\value{
A rezrDF object with the added / changed field.
}
\description{
Easily add a field to / change a field in a rezrDF using information from another rezrDF
}
