% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj_easy.R
\name{addFieldForeign.rezrObj}
\alias{addFieldForeign.rezrObj}
\title{Easily add a field to / change a field in a rezrDF using information from another rezrDF}
\usage{
\method{addFieldForeign}{rezrObj}(
  rezrObj,
  targetEntity,
  targetLayer = "",
  sourceEntity,
  sourceLayer = "",
  targetForeignKeyName,
  targetFieldName = "",
  sourceFieldName = "",
  type = "simple",
  fieldaccess = "flex",
  complexAction = NULL
)
}
\arguments{
\item{rezrObj}{The Rezonator object that you will be changing.}

\item{targetEntity}{The Rezontor entity that you will be changing (e.g. token, unit, entry ...)}

\item{targetLayer}{The layer of the Rezonator entity that you will be changing from. Leave blank for entities without layers (i.e. token, entry and unit). If you don't have layers for other entities, type 'default'.}

\item{sourceEntity}{The Rezonator entity that you will be getting your information from.}

\item{sourceLayer}{The layer of the Rezonator entity that you will be getting your information from. Leave blank for entities without layers (i.e. token, entry and unit). If you don't have layers for other entities, type 'default'.}

\item{targetForeignKeyName}{The name of the field in the target DF that corresponds to the key (ID) of the source DF. If type = "simple", this is a column in the rezrDF. If type = "complex", this is a field in the target node map, and targetNodeMap must be specified.}

\item{targetFieldName}{The name of the field you want to add / change.}

\item{sourceFieldName}{The name of the field in the source rezrDF that you will be pulling information from.}

\item{type}{If "simple", each row of the target DF gets information from only one row of the source DF. If "complex", each row of the target DF gets information from multiple rows of the source DF.}

\item{fieldaccess}{The field access value you want to set to the new / changed field.}

\item{complexAction}{When type = "complex", the action that will aggregate the values from the source DF together. See \link{complexActions} for ideas.}
}
\value{

}
\description{
Easily add a field to / change a field in a rezrDF using information from another rezrDF
}
