% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj_easy.R
\name{addFieldLocal.rezrObj}
\alias{addFieldLocal.rezrObj}
\alias{changeFieldLocal.rezrObj}
\title{Easily add a field to / change a field in a rezrDF from a rezrObj using only information from that rezrDF}
\usage{
\method{addFieldLocal}{rezrObj}(
  rezrObj,
  entity,
  layer,
  fieldName,
  expression,
  fieldaccess = "flex"
)

\method{changeFieldLocal}{rezrObj}(
  rezrObj,
  entity,
  layer,
  fieldName,
  expression,
  fieldaccess = "flex"
)
}
\arguments{
\item{rezrObj}{The rezrObj that you will be changing.}

\item{entity}{The name of the Rezonator entity that you will be changing - track, rez, chunk, token, unit, stack, etc.}

\item{layer}{The name of the layer of the Rezonator entity that you wil be changing. Leave blank for entities without layers (i.e. token, entry and unit). If you don't have layers for other entities, type 'default'.}

\item{fieldName}{The name of the field you would like to add or change.}

\item{expression}{An R expression. For example, if you would like to add the columns a and b of the rezrDF together, type a + b.}

\item{fieldaccess}{The access value that you would like to assign to the new or changed field.}
}
\value{
A rezrObj with the changed rezrDF.
}
\description{
Easily add a field to / change a field in a rezrDF from a rezrObj using only information from that rezrDF
}
