% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/time_seq.R
\name{addIsWordField}
\alias{addIsWordField}
\alias{addIsWordField.rezrDF}
\alias{addIsWordField.rezrObj}
\title{Add a field on whether something is a word or not.}
\usage{
addIsWordField(x, ...)

\method{addIsWordField}{rezrDF}(x, cond, addWordSeq = T)

\method{addIsWordField}{rezrObj}(x, cond, addWordSeq = T)
}
\arguments{
\item{x}{The rezrDF or rezrObj to be edited.}

\item{cond}{The wordhood condition. For example, if your word column is called 'word', and you wish to exclude zeroes, you may write 'x == "<0>"'.}

\item{addWordSeq}{If TRUE, the columns wordOrder and docWordSeq will be added.}
}
\value{
The modified rezrDF or rezrObj. If addWordSeq is set to TRUE, the columns wordOrder and docWordSeq will be added to tokenDF and entryDF, and the columns wordOrderFirst, wordOrderLast, docWordSeqFirst and docWordSeqLast will be added to unitDF, chunkDF and trackDF.
}
\description{
Add a field on whether something is a word or not.
}
\note{
If used on a rezrObj and addWordSeq = T, wordOrder and docWordSeq are automatically added to entry, unit, chunk and track tables. Rez and stack tables coming soon.
}
