% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{addPositionAmongSiblings}
\alias{addPositionAmongSiblings}
\title{Find the position of a chunk among its siblings in a tree.}
\usage{
addPositionAmongSiblings(
  chunkDF,
  rezrObj,
  treeEntryDFAddress = NULL,
  cond = expr(TRUE)
)
}
\arguments{
\item{chunkDF}{The \code{chunkDF} to be edited.}

\item{rezrObj}{The \code{rezrObj} containing the \code{chunkDF}.}

\item{treeEntryDFAddress}{The address to the \code{treeEntryDF} layer that will be used. If left blank, all layers will be used.}

\item{cond}{A condition for excluding siblings (for example, adjuncts and auxiliaries).}
}
\value{
The \code{chunkDF} with a column \code{siblingPos} giving the position of the chunk among its siblings in the tree.
}
\description{
Find the position of a chunk among its siblings in a tree.
}
\examples{
sbc007 = getAllTreeCorrespondences(sbc007)
sbc007$chunkDF$refexpr = addPositionAmongSiblings(sbc007$chunkDF$refexpr, sbc007)
}
