% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_seq.R
\name{addUnitSeq}
\alias{addUnitSeq}
\title{Add unit sequence information to different levels of rezrObj.}
\usage{
addUnitSeq(rezrObj, entity, layers = "")
}
\arguments{
\item{rezrObj}{The rezrObj object}

\item{entity}{The name of the Rezonator entity that you will be changing - track, rez, chunk, token, unit, stack, etc.}

\item{layers}{The layers to which unit sequence information is to be added.}
}
\value{
The rezrObj object with unit sequences in the entity desired, plus all levels below. For example, if your entity is 'track', you will see unitSeq information on token and chunk too, but not rez.
}
\description{
Add unit sequence information to different levels of rezrObj.
}
