% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rez.R
\name{augmentRezInfo}
\alias{augmentRezInfo}
\title{Augment unitDF and cliqueDF with resonance-related information.}
\usage{
augmentRezInfo(rezrObj)
}
\arguments{
\item{rezrObj}{The \code{rezrObj} to be augmented.}
}
\value{
A \code{rezrObj} with the \code{unitDF} updated to contain information about the level of resonance that unit is involved in, and \code{cliqueDF} updated to contain information on whether the clique contains other-resonance.
}
\description{
Augment unitDF and cliqueDF with resonance-related information.
}
