% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj.R
\name{combineLayers}
\alias{combineLayers}
\alias{combineChunks}
\alias{combineTokenChunk}
\title{Combine different layers of an entity, and similar entities}
\usage{
combineLayers(rezrObj, entity, type = "intersect")

combineChunks(rezrObj, ...)

combineTokenChunk(rezrObj, type = "intersect")
}
\arguments{
\item{rezrObj}{A rezrObj object}

\item{entity}{The Rezonator entity with multiple layers (e.g. track, rez, chunk) that you want to combine}

\item{type}{Do you want the resultant rezrDF to contain the intersection of the columns in all the components, or the union (with absent fields becoming NA)?}
}
\value{
A rezrDF containing the required material
}
\description{
Combine different layers of an entity, and similar entities
}
