% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_lowerToHigher.R
\name{concatStringFields}
\alias{concatStringFields}
\title{Concatenate string values in a lower-level data frame}
\usage{
concatStringFields(
  simpleDF,
  complexDF,
  complexNodeMap,
  fieldnames,
  tokenListName = "tokenList",
  separator = "",
  ...
)
}
\arguments{
\item{simpleDF}{The lower-level dataframe, for example the token dataframe for chunks and units, or the unit dataframe for stacks.}

\item{complexDF}{The dataframe that you're trying to add the concatenated fields to.}

\item{complexNodeMap}{The node map corresponding to the simpleDF.}

\item{fieldnames}{The fields to be concatenated.}

\item{separator}{The character by which words will be separated.}

\item{...}{Additional fields 'simpleDFAddress', 'complexNodeMapAddress', 'fieldaccess' (foreign by default) from \code{\link[=lowerToHigher]{lowerToHigher()}}. Only needed if you want automatically generated update functions.}
}
\value{
complexDF
}
\description{
Not recommended to be called by most users; \link{addFieldForeign} with \link{concatenateAll} suffices for most purposes.
}
