% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{createLowerToHigherUpdate}
\alias{createLowerToHigherUpdate}
\title{Create an update function based on a lowerToHigher-type action.}
\usage{
createLowerToHigherUpdate(
  df,
  rezrObj,
  address,
  fkeyAddress,
  action,
  field = "",
  fkeyInDF = FALSE,
  seqName = "discourseTokenSeq"
)
}
\arguments{
\item{df}{The target rezrDF you want to modify.}

\item{rezrObj}{The rezrObj.}

\item{address}{The address to the info field that you want from your source DF.}

\item{fkeyAddress}{The address to the token list in the nodeMap that corresponds to the target rezrDF.}

\item{seqName}{}
}
\value{

}
\description{
Create an update function based on a lowerToHigher-type action.
}
