% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createRezId}
\alias{createRezId}
\title{Create a Rezonator ID.}
\usage{
createRezId(n = 1, existing = character(0))
}
\arguments{
\item{n}{The number of Rezonator IDs you want.}

\item{existing}{Existing IDs, to avoid overlap.}
}
\value{
A vector of Rezonator IDs.
}
\description{
Create a Rezonator ID.
}
\examples{
createRezId(3)
}
