% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_mutate.R
\name{createUpdateFunction}
\alias{createUpdateFunction}
\title{Create an update function.}
\usage{
createUpdateFunction(field, x, df)
}
\arguments{
\item{field}{The field for which you want to create an update function.}

\item{x}{An R expression. For example, if you want to column2 to be updated to always be three times column3, then x should be column3 * 3.}

\item{df}{The rezrDF for which you want to create an update function.}
}
\value{
An update function with automatically generated dependency information. I will figure out the dependency information for you, so you don't have to define it yourself.
}
\description{
This is for 'auto' fields only; 'foreign' field take createUpdateFunction.
}
