% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{deps}
\alias{deps}
\alias{deps<-}
\title{Extract/set an attribute of a updateFunction.}
\usage{
deps(updateFunct)

deps(updateFunct) <- value
}
\arguments{
\item{updateFunct}{An updateFunction object.}

\item{value}{The value you want to set the function to.}
}
\description{
Extract dependency information from an updateFunction, or set it.
}
