% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{addFieldLocal}
\alias{addFieldLocal}
\alias{changeFieldLocal}
\alias{addFieldForeign}
\alias{changeFieldForeign}
\title{Generic functions for easy editing}
\usage{
addFieldLocal(x, ...)

changeFieldLocal(x, ...)

addFieldForeign(x, ...)

changeFieldForeign(x, ...)
}
\arguments{
\item{x}{The object to be modified.}

\item{...}{Argumnts to be passed to the specific functions.}
}
\value{

}
\description{
Generic functions for easy editing
}
\note{
For details of these functions, see \link{addFieldLocal.rezrDF} / \link{addFieldForeign.rezrDF} for rezrDFs and \link{addFieldLocal.rezrObj} / \link{addField.rezrObjForeign} for rezrObjs.
}
