% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{findContainingChunk}
\alias{findContainingChunk}
\title{Find chunks containing other chunks}
\usage{
findContainingChunk(containedDF, containerDF, proper = F)
}
\arguments{
\item{containedDF}{The rezrDF with contained chunks (may be a track/rez rezrDF)}

\item{containerDF}{The rezrDF with container chunks (may be a track/rez rezrDF)}

\item{proper}{Do you count only proper containment?}
}
\value{
A vector of container chunks, each of which corresponds to one row in the containedDF (the first one is chosen in the case of multiple parents). Right now, we do not support criteria for choosing container chunk.
}
\description{
Find chunks containing other chunks
}
\examples{
sbc007$chunkDF$verb = sbc007$chunkDF$verb \%>\% rez_mutate(containingRefexpr = findContainingChunk(sbc007$chunkDF$verb, sbc007$chunkDF$refexpr, proper = T))
}
