% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rez.R
\name{findResonancesBetween}
\alias{findResonancesBetween}
\title{Find resonances.between two sets of units.}
\usage{
findResonancesBetween(rezrObj, units_1, units_2, rez_layer = "default")
}
\arguments{
\item{rezrObj}{The rezrObj.}

\item{units_1}{A list of vectors of units.}

\item{units_2}{A list of vectors of units, same length as units_1.}

\item{rez_layer}{The resonance layer you're considering.}
}
\value{
A vector containing the number of resonances between the first vector of units_1 and first vector of units_2, second vector of units_1 and second vector of units_2, etc.
}
\description{
Find resonances.between two sets of units.
}
