% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{getAllTreeCorrespondences}
\alias{getAllTreeCorrespondences}
\title{Relate table rows to tree entries covering the same tokens.}
\usage{
getAllTreeCorrespondences(rezrObj, entity = "chunk")
}
\arguments{
\item{rezrObj}{The rezrObj object}

\item{entity}{The name of the Rezonator entity that you will be changing - track, rez, chunk, token, unit, stack, etc.}
}
\value{
The rezrObj with a new column called 'treeEntry' added to rows that correspond to tree entries. If called on 'track', treeEntry is added to token and chunk as well. Not available for units.
}
\description{
Relate table rows to tree entries covering the same tokens.
}
\examples{
sbc007 = getAllTreeCorrespondences(sbc007, entity = "track")
}
