% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_seq.R
\name{getBiluoFromOrder}
\alias{getBiluoFromOrder}
\title{Get BILUO (begin-intermediate-last-unique-outside) values from a sequence value.}
\usage{
getBiluoFromOrder(seq, length)
}
\arguments{
\item{seq}{The sequence number of the entity within a larger structure.
For example, \code{tokenOrder} for the position of a token within a unit.}

\item{length}{The length of the sequence, which you can get from \code{inLength()}}
}
\value{
A vector of BILUO values corresponding to the sequence values. B = beginning of the larger structure,
L = last element of the larger structure, I = intermediate element of the larger structure,
U = only element of the larger structure, O = not within the larger structure.
}
\description{
Get BILUO (begin-intermediate-last-unique-outside) values from a sequence value.
}
\examples{
sbc007$tokenDF = addFieldLocal(sbc007$tokenDF, "pSentLength", inLength(pSentOrder, isWord = (text != "<0>")), type = "complex", groupField = "pSent")
sbc007$tokenDF = changeFieldLocal(sbc007$tokenDF, "pSentBiluo", getBiluoFromOrder(pSentOrder, pSentLength))
}
