% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{getGantt}
\alias{getGantt}
\title{Get Gantt charts from rezrObj.}
\usage{
getGantt(
  rezrObj,
  x = "token",
  y = "participant",
  obj = "unit",
  thickness = 12,
  stacking = "Turn"
)
}
\arguments{
\item{rezrObj}{The rezrObj that the Gantt chart is to be created fromm}

\item{x}{The unit of the x-axis (token, unit ...)}

\item{y}{The unit of the y-axis (generally participant, I don't support anything else at the moment)}

\item{obj}{What does each rectangle on the Gantt chart represent? By default, "unit", but can also be "stack".}

\item{thickness}{Thickness of the rectangles.}

\item{stacking}{The stacking that you would like the rectangles to be, if obj = "stack".}
}
\description{
Get Gantt charts from rezrObj.
}
