% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{getPosInChain}
\alias{getPosInChain}
\title{Get the position of a chain entry (track or rez) in a chain (trail or resonance).}
\usage{
getPosInChain(
  tokenOrder = NULL,
  chain = NULL,
  exclFrag = F,
  combinedChunk = NULL,
  nonFragmentMember = F
)
}
\arguments{
\item{tokenOrder}{The vector of sequence values values where the mentions appeared. Common choices are docTokenSeqFirst, docTokenSeqLast, wordTokenSeqFirst and wordTokenseqLast (the last two are available after running \link{addIsWordField} on a rezrObj. By default it's docTokenSeqLast.}

\item{chain}{The chain that each mention belongs to.}

\item{exclFrag}{Exclude 'fragments' (i.e. members of a combined chunk which do not serve as meaningful chunks in their own right)}

\item{combinedChunk}{The \code{combinedChunk} column of the rezrDF. By default, named \code{combinedChunk}.}

\item{nonFragmentMember}{Vector indicating whether each entry is a non-fragment member, i.e. a member of a combined chunk that also serves as a meaningful chunk in its own right.}
}
\value{
The
}
\description{
Get the position of a chain entry (track or rez) in a chain (trail or resonance).
}
