% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{getRezrDFAttr}
\alias{getRezrDFAttr}
\alias{fieldaccess}
\alias{updateFunct}
\alias{inNodeMap}
\alias{setRezrDFAttr}
\alias{updateFunct<-}
\alias{fieldaccess<-}
\alias{inNodeMap<-}
\title{Extract/set an attribute of a rezrDF.}
\usage{
getRezrDFAttr(df, attr, fields = "")

fieldaccess(df, fields = "")

updateFunct(df, fields = "")

inNodeMap(df, fields = "")

setRezrDFAttr(df, attr, fields = "", value)

updateFunct(df, fields = "") <- value

fieldaccess(df, fields = "") <- value

inNodeMap(df, fields = "") <- value
}
\arguments{
\item{df}{The \code{rezrDF} whose field access attributes you want to see.}

\item{attr}{The attribute that you want to extract.}

\item{fields}{The field whose access attribute you want to see. If left blank, a vector containing all the attributes is output.}

\item{value}{The value that you want to set it to.}
}
\value{
The value of the attribute you want to extract (for extract functions), and the DF with the value set (for set functions).
}
\description{
Extract an attribute of a data.frame (\code{inNodeMap}, \code{fieldaccess}, \code{updateFunct}), or set it.
}
\examples{
updateFunct(sbc007$unitDF, "docTokenSeqLast")
#If you want to be able to edit the timestamps manually
fieldaccess(sbc007$unitDF, c("unitEnd", "unitStart")) = c("flex", "flex")
inNodeMap(sbc007$unitDF)
}
