% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_lowerToHigher.R
\name{getSeqBounds}
\alias{getSeqBounds}
\title{Get the max and min of a certain value (typically sequence or time) from a lower object table to a higher object table.}
\usage{
getSeqBounds(
  simpleDF,
  complexDF,
  complexNodeMap,
  fieldnames,
  simpleIsAtom = T,
  seqName = "",
  tokenListName = "tokenList",
  ...
)
}
\arguments{
\item{simpleDF}{The lower-level dataframe, for example the token dataframe for chunks and units, or the unit dataframe for stacks. There should either be an integer field that contains the sequence in question, or two integer fields containing the first and last (see simpleIsAtom).}

\item{complexDF}{The dataframe that you're trying to add the concatenated fields to. We will create two integer fields, one for the first and one for the last integer.}

\item{complexNodeMap}{The node map corresponding to the simpleDF.}

\item{fieldnames}{The fields to be concatenated.}

\item{simpleIsAtom}{If set to T, that means the simpleDF only contains single values, not a range. If set to F, that means the simpleDF contains a range, i.e. somethingFirst and somethingLast.}

\item{...}{Additional fields 'simpleDFAddress', 'complexNodeMapAddress', 'fieldaccess' (foreign by default) from \code{\link[=lowerToHigher]{lowerToHigher()}}. Only needed if you want automatically generated update functions.}
}
\value{
A complex DF.
}
\description{
Get the max and min of a certain value (typically sequence or time) from a lower object table to a higher object table.
}
