% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_seq.R
\name{getSeqFromID}
\alias{getSeqFromID}
\title{Get sequence number (position) within a larger structure (turn, prosodic sentence, etc.) from the ID of that structure}
\usage{
getSeqFromID(id)
}
\arguments{
\item{id}{The ID of the larger structure within which you would like to find the position of an individual component.
For example, if you want to find the position of a word within a prosodic sentence from the prosodic sentence ID.}
}
\value{
A vector of sequence values within that structure. For example, if you are working with the tokenDF and \code{id}
gives the prosodic sentence ID, then this function will return the position of a token within a prosodic sentence.
}
\description{
Get sequence number (position) within a larger structure (turn, prosodic sentence, etc.) from the ID of that structure
}
