% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_tidy.R
\name{rez_group_by}
\alias{rez_group_by}
\alias{rez_ungroup}
\title{Group / ungroup rezrDFs.}
\usage{
rez_group_by(df, ...)

rez_ungroup(df, ...)
}
\arguments{
\item{df}{The data frame to be modified.}

\item{...}{Arguments to be passed to group_by / ungroup, including the field(s) on which you're grouping.}
}
\value{
The grouped / ungroupd rezrDF.
}
\description{
A replacement for dplyr group_by and ungroup.
}
