% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRez.R
\name{importRez}
\alias{importRez}
\title{Import a Rez file}
\usage{
importRez(path)
}
\arguments{
\item{path}{}
}
\value{
rezRObject
}
\description{
Import a Rez file, returning an R object containing a the node map, word grid, unit grid and chain grid.
}
