% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_lowerToHigher.R
\name{lowerToHigher}
\alias{lowerToHigher}
\title{Combine information from multiple entries in a lower-level source table to a higher-level target table.}
\usage{
lowerToHigher(
  simpleDF = NULL,
  complexDF,
  complexNodeMap = NULL,
  fieldnames,
  higherFieldnames = "",
  action,
  seqName = "discourseTokenSeq",
  tokenListName = "tokenList",
  simpleDFAddress = "",
  complexNodeMapAddress = "",
  rezrObj = NULL,
  fieldaccess = "foreign"
)
}
\arguments{
\item{simpleDF}{The DF containing simple information (source table).}

\item{complexDF}{The DF to contain complex information from the simple table (target table).}

\item{complexNodeMap}{The node map corresponding to the complex (target) table.}

\item{fieldnames}{The field names in the simple table whose information is to be extracted.}

\item{higherFieldnames}{The field names in the complex table whose information is to be extracted.}

\item{action}{The action to be performed on the data from the source table.}

\item{seqName}{The name of the sequence field in the source table.}

\item{tokenListName}{The name of the list of foreign keys in the target table.}

\item{simpleDFAddress}{An address to the simple (source) DF. Only needed if you want an automatic update function.}

\item{complexNodeMapAddress}{An address to the complex (target) DF. Only needed if you want an automatic update function.}

\item{rezrObj}{A rezrObj object.}

\item{fieldaccess}{The field access value of the field. If not set to foreign, no update function will be automatically added.}
}
\value{
The modified target table (complexDF).
}
\description{
Combine information from multiple entries in a lower-level source table to a higher-level target table.
}
