% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_easy.R
\name{mergeCats}
\alias{mergeCats}
\alias{renameCats}
\title{Merging and renaming categories}
\usage{
mergeCats(x, ..., asFactor = F, levels = NULL)

renameCats(x, ..., asFactor = F, levels = NULL)
}
\arguments{
\item{x}{A column or vector whose categories are to be merged or renamed.}

\item{...}{The name of each argument is a new category, and the value of each argument is a vector of names of old categories (as character values, even if the original column/vector contains factors).}

\item{asFactor}{Do you want the result to be a factor?}

\item{levels}{If asFactor = T, you an use this to set the levels of the factor.}
}
\value{
A column or vector with the desired changes.
}
\description{
Functions for merging and renaming categories in character or factor columns.
}
