% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{mergeChunksWithIDs}
\alias{mergeChunksWithIDs}
\title{Relate table rows to tree entries covering the same tokens.}
\usage{
mergeChunksWithIDs(rezrObj, idField, addToTrack = F, selectCond = NULL)
}
\arguments{
\item{rezrObj}{The rezrObj object}

\item{addToTrack}{Do you want to add the chunks to the trackDF as well?}

\item{selectCond}{The condition for selecting which chunk provides the field values.}

\item{treeEntryDF}{A treeEntry data.frame, possibly filtered.}
}
\value{
The rezrObj with additional rows for merged chunks. Original chunks stay behind. There will be a new column called combinedChunk. Combined chunks will get the value 'combined'. Members of those chunks will get the value '|infomember=COMBINEDCHUNKID' (if they are providing the data for the combined chunk) or '|member=COMBINEDCHUNKID' (if they are not the data-providing chunk). treeEntry (through getAllTreeCorrespondences()) is required to be present in the chunkDF.
}
\description{
Relate table rows to tree entries covering the same tokens.
}
\examples{
sbc007 = mergeChunksWithIDs(sbc007, "largerChunk", selectCond = NULL)
}
