% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{mergeChunksWithTree}
\alias{mergeChunksWithTree}
\title{Relate table rows to tree entries covering the same tokens.}
\usage{
mergeChunksWithTree(
  rezrObj,
  treeEntryDF = NULL,
  addToTrack = F,
  selectCond = NULL
)
}
\arguments{
\item{rezrObj}{The rezrObj object}

\item{treeEntryDF}{A treeEntry data.frame, possibly filtered.}

\item{addToTrack}{Do you want to add the chunks to the trackDF as well?}

\item{selectCond}{The condition for selecting which chunk provides the field values.}
}
\value{
The rezrObj with additional rows for merged chunks. Original chunks stay behind.
}
\description{
Relate table rows to tree entries covering the same tokens.
}
