% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{new_rezrDF}
\alias{new_rezrDF}
\title{Constructor function for rezrDF}
\usage{
new_rezrDF(df, fieldaccess, updateFunct, inNodeMap)
}
\arguments{
\item{df}{The data.frame}

\item{fieldaccess}{A set of field access values}

\item{updateFunct}{A set of update functions}

\item{inNodeMap}{Whether the fields of the data.frame are in the node map}
}
\value{
A rezrDF object with the fieldaccess, updateFunct and inNodeMap values provided.
}
\description{
Constructor function for rezrDF
}
\note{
Field access values include 'key' for the primary key, 'core' fields for fields mostly generated in Rezonator, 'flex' for fields that may be flexibly changed, 'auto' for fields that are automatically generated from information from the SAME table, and 'foreign' for fields that are automatically generated from information from  ANOTHER table.
}
