% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_seq.R
\name{getOrderFromSeq}
\alias{getOrderFromSeq}
\alias{getSeqFromOrder}
\title{Get sequence number (position) within a larger structure (turn, prosodic sentence, etc.) from the ID of that structure, or the other way around}
\usage{
getOrderFromSeq(id)

getSeqFromOrder(order)
}
\arguments{
\item{id}{The ID of the larger structure within which you would like to find the position of an individual component.
For example, if you want to find the position of a word within a prosodic sentence from the prosodic sentence ID.}

\item{order}{The position of an individual component within a larger structure.}
}
\value{
A vector of sequence values within that structure. For example, if you are working with the tokenDF and \code{id}
gives the prosodic sentence ID, then this function will return the position of a token within a prosodic sentence.

A vector of either order values within the structure . For example, if you are working with the \code{tokenDF} and \code{id}
gives the prosodic sentence ID, then \code{getOrderFormSeq} will return the position of a token within a prosodic sentence. If \code{order} gives the position within the prosodic sentence, then \code{getSeqFromOrder} will return the prosodic sentence sequence.
}
\description{
Get sequence number (position) within a larger structure (turn, prosodic sentence, etc.) from the ID of that structure, or the other way around
}
