% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_bridge.R
\name{reducedFrameMatrix}
\alias{reducedFrameMatrix}
\title{Extract a frameMatrix, removing rows and columns that do not participate in frame semantics.}
\usage{
reducedFrameMatrix(rezrObj, cond)
}
\arguments{
\item{rezrObj}{The \code{rezrObj} from which the \code{frameMatrix} is extracted.}

\item{cond}{Conditions for reduction.}
}
\description{
Extract a frameMatrix, removing rows and columns that do not participate in frame semantics.
}
