% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rezrDF.R
\name{reload}
\alias{reload}
\alias{reload.rezrDF}
\alias{reloadLocal}
\alias{reloadForeign}
\title{Functions for reloading auto and foreign fields.}
\usage{
reload(x, ...)

\method{reload}{rezrDF}(df, rezrObj, fields = "")

reloadLocal(df, fields = "")

reloadForeign(df, rezrObj, fields = "")
}
\arguments{
\item{df}{The rezrDF you want to modify.}

\item{rezrObj}{The entire rezrObj.}

\item{fields}{The fields of the rezrDF you would like to modify.}
}
\value{
The modified rezrDF. reloadLocal reloads only auto fields; reloadForeign reloads only foreign fields. Foreign fields are always updated before auto fields. If the fields parameter is not available, I will find the best update order; otherwise, I will use the order you provide except for the rule where foreign fields are updated first.
}
\description{
Functions for reloading auto and foreign fields.
}
\note{
reload.rezrObj is planned for a future release, but not yet available
}
\examples{
sbc007$tokenDF = changeFieldLocal(sbc007$tokenDF, fieldName = "text", expression = case_when(text == "<0>" ~ "∅", T ~ text))
sbc007$entryDF = sbc007$entryDF \%>\% reloadForeign(sbc007)
sbc007$unitDF = sbc007$unitDF \%>\% reload(sbc007)
sbc007$unitDF \%>\% filter(str_detect(text, "∅")) \%>\% rez_select(id, text) \%>\% head
}
