% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{reloadForeign}
\alias{reloadForeign}
\title{Reload foreign fields in a rezrDF.}
\usage{
reloadForeign(df, rezrObj, fields = "")
}
\arguments{
\item{df}{The rezrDF you want to modify.}

\item{rezrObj}{The entire rezrObj.}

\item{fields}{The fields of the rezrDF you want to modify.}
}
\value{
The modified rezrDF.
}
\description{
Reload foreign fields in a rezrDF.
}
