% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{rez_group_by}
\alias{rez_group_by}
\title{Group rezrDFs together.}
\usage{
rez_group_by(df, ...)
}
\arguments{
\item{...}{Arguments to be passed to group_by, including the field(s) on which you're grouping.}
}
\value{

}
\description{
A replacement for dplyr group_by.
}
