% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlineaR-rezonateR.R
\name{rez_pangloss}
\alias{rez_pangloss}
\title{Read Pangloss file}
\usage{
rez_pangloss(url, filename)
}
\arguments{
\item{url}{A URL for the Pangloss source.}

\item{filename}{a filename for the export.}
}
\value{
rezImport
}
\description{
Read Pangloss file
}
\examples{
rez_pangloss("https://cocoon.huma-num.fr/data/mazaudon/masters/crdo-TAJ_ORIGTAM.xml", "tamang1.txt")
}
