% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{rez_rbind}
\alias{rez_rbind}
\title{Bind together related rezrDF objects.}
\usage{
rez_rbind(..., type = "intersect")
}
\arguments{
\item{...}{The rezrDF objects to be combined}

\item{type}{The type of combination. If 'intersect', I will take the intersection of the columns in the rezrDFs. If 'union', I will take the union of the columns, populating missing fields with NAs.}
}
\value{
The bound rezrDF
}
\description{
Bind together related rezrDF objects.
}
