% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_tidy.R
\name{rez_rename}
\alias{rez_rename}
\title{Rename rezrDF columns.}
\usage{
rez_rename(df, ...)
}
\arguments{
\item{df}{The data frame to be modified.}

\item{...}{Functions to be passed to rename. New column names are argument names; old column names are argument values.}
}
\value{
A rezrDF object.
}
\description{
Rename rezrDF columns.
}
\note{
This function does not update foreign references to the field that you're renaming. So be sure to update the updateFunctions of those fields; otherwise, you will break your rezrObj.
}
