% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{rez_select}
\alias{rez_select}
\title{Select columns in a rezrDF}
\usage{
rez_select(df, ...)
}
\arguments{
\item{df}{A rezrDF object}

\item{...}{Functions to be passed to dplyr select}
}
\value{
The rezrDF object with only the required columns
}
\description{
Select columns in a rezrDF
}
