% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_mutate.R
\name{rez_validate_fieldchange}
\alias{rez_validate_fieldchange}
\title{Validate a field change.}
\usage{
rez_validate_fieldchange(df, changedFields)
}
\arguments{
\item{df}{The rezrDF you are planning to change.}

\item{changedFields}{The fields you want to change.}
}
\description{
This function ensures that the fields in a DF you wish to change are actually a good idea to change. It produces an error if a primary key is among the fields you wish to change, and warnings otherwise. This is automatically called if you use a rezonateR-internal function such as rez_mutate.
}
