% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_csv.R
\name{rez_write_csv}
\alias{rez_write_csv}
\title{Read and write rezrDFs as CSV files}
\usage{
rez_write_csv(df, path, inclCols = character(0), exclCols = character(0), ...)
}
\arguments{
\item{df}{The rezrDF to be exported.}

\item{path}{The path from which a CSV is to be imported / exported.}

\item{inclCols}{Columns to be included from the import / export.}

\item{exclCols}{Columns to be excluded from the import / export.}

\item{...}{Arguments passed onto read_csv and write_csv.}

\item{lubridate}{Are you using the lubridate package?}

\item{origDF}{The rezrDF that originally produced this CSV, used to identify data types of the columns.}
}
\value{
For rez_read_csv, the CSV being imported.
}
\description{
Read and write rezrDFs as CSV files
}
