% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rezrDF_easy.R, R/rezrObj_easy.R
\name{removeField}
\alias{removeField}
\alias{removeField.rezrDF}
\alias{removeField.rezrObj}
\title{Remove a field from a rezrDF.}
\usage{
removeField(x, ...)

\method{removeField}{rezrDF}(rezrDF, fields)

\method{removeField}{rezrObj}(rezrObj, entity, layer, fields)
}
\arguments{
\item{fields}{Vector of column names to be removed.}
}
\value{
A rezrDF with the specified fields removed.
}
\description{
Remove a field from a rezrDF.
}
