% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj.R
\name{rez_save}
\alias{rez_save}
\alias{rez_load}
\title{Save and load rezrObj and rezrDF objects.}
\usage{
rez_save(obj, filename)

rez_load(filename)
}
\arguments{
\item{obj}{The object you would like to save.}

\item{filename}{The filename, ending in .Rdata, you would like to save to.}
}
\value{
For rez_load, the saved rezrObj or rezrDF.
}
\description{
Save and load rezrObj and rezrDF objects.
}
