% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbc007}
\alias{sbc007}
\title{A Tree's Life - preliminary annotation}
\format{
A \code{rezrObj} with the following elements: \code{nodeMap}, \code{chunkDF}, \code{docDF}, \code{entryDF}, \code{linkDF}, \code{mergedDF}, \code{resonanceDF}, \code{rezDF},
\code{tokenDF}, \code{trackDF}, \code{trailDF}, \code{treeDF}, \code{treeEntryDF}, \code{treeLinkDF}, \code{unitDF}
}
\usage{
sbc007
}
\description{
A \code{rezrObj} that comes from a preliminary annotation of part of the document 'A Tree's Life' from the
Santa Barbara Corpus of Spoken American English (SBC). It has been annotated for resonance and coreference.
Coreference chains include referential expressions along with non-referring expressions that have been otherwise
referred back to (e.g. discourse deixis). Two-layered trees indicate argument structure of each verb. The verbal complex (including
auxiliaries and adverbs that intervene between the auxiliary and verb, but not arguments that intervene between auxiliary
and verb) is annotated as blank chunks (with \code{chunkType = "verb"}) and serves as the root of each tree. Tree links are annotated as \code{Relation = "Subject"}
when it indicates a subject-verb relation; otherwise \code{Relation} is left blank.
}
\keyword{datasets}
