% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{stringToFactor}
\alias{stringToFactor}
\title{Change the strings in a rezrDF to factors}
\usage{
stringToFactor(df, colsToChange = NULL, levels = list())
}
\arguments{
\item{df}{The data.frame to be modified.}

\item{colsToChange}{The columns to be modified. If left blank, all character columns will be factor-fied.}

\item{levels}{Factor levels, as a list with each label being a column name.}
}
\description{
Change the strings in a rezrDF to factors
}
