% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{lastMentionUnit}
\alias{lastMentionUnit}
\alias{unitsToLastMention}
\alias{lastMentionToken}
\alias{tokensToLastMention}
\alias{countPrevMentions}
\alias{countPrevMentionsIf}
\alias{countPrevMentionsMatch}
\alias{getPrevMentionField}
\alias{nextMentionUnit}
\alias{unitsToNextMention}
\alias{nextMentionToken}
\alias{tokensToNextMention}
\alias{countNextMentions}
\alias{countNextMentionsIf}
\alias{countNextMentionsMatch}
\alias{getNextMentionField}
\title{Functions related to mentions of the same entity in previous/following context in track chains.}
\usage{
lastMentionUnit(unitSeq = NULL, chain = NULL)

unitsToLastMention(unitSeq = NULL, chain = NULL)

lastMentionToken(tokenSeq = NULL, chain = NULL)

tokensToLastMention(
  tokenSeq = NULL,
  chain = NULL,
  zeroProtocol = "literal",
  zeroCond = NULL,
  unitSeq = NULL,
  unitDF = NULL
)

countPrevMentions(windowSize, unitSeq = NULL, chain = NULL)

countPrevMentionsIf(windowSize, cond, unitSeq = NULL, chain = NULL)

countPrevMentionsMatch(windowSize, field, unitSeq = NULL, chain = NULL)

getPrevMentionField(field, tokenSeq = NULL, chain = NULL)

nextMentionUnit(unitSeq = NULL, chain = NULL)

unitsToNextMention(unitSeq = NULL, chain = NULL)

nextMentionToken(tokenSeq = NULL, chain = NULL)

tokensToNextMention(
  tokenSeq = NULL,
  chain = NULL,
  zeroProtocol = "literal",
  zeroCond = NULL,
  unitSeq = NULL,
  unitDF = NULL
)

countNextMentions(windowSize, unitSeq = NULL, chain = NULL)

countNextMentionsIf(windowSize, cond, unitSeq = NULL, chain = NULL)

countNextMentionsMatch(windowSize, field, unitSeq = NULL, chain = NULL)

getNextMentionField(field, tokenSeq = NULL, chain = NULL)
}
\arguments{
\item{unitSeq}{The vector of tokenSeq values where the mentions appeared. You can choose tokenSeqFirst, tokenSeqFirst, or maybe an average of the two. By default it's tokenSeqFirst.}

\item{chain}{The chain that each mention belongs to.}

\item{tokenSeq}{The vector of sequence values values where the mentions appeared. Common choices are discourseTokenSeqFirst, discourseTokenSeqLast, wordTokenSeqFirst and wordTokenseqLast (the last two are available after running \link{addIsWordField} on a rezrObj. By default it's discourseTokenSeqLast.}

\item{zeroProtocol}{If 'literal', I will take the seq values of the zeroes at face value. (If you set zeros as non-words and use discourseWordSeqFirst or discourseWordSeLast as your tokenSeq, this will lead to meaningless values.) If 'unitFinal', I will treat zeroes as if they were the final word of the unit. If 'unitFirst', I will treat zeroes as if they were the first word of the unit.}

\item{zeroCond}{A condition for determining whether a token is zero. For most people, this should be (word column) == "<0>".}

\item{windowSize}{The size of the window in which you will be counting.}

\item{cond}{For if functions, the condition that the previous / next mention must satisfy. It cannot refer to the current mention.}

\item{field}{The field whose value you want to match or extract.}
}
\description{
See also \link{countCompetitors}.
}
