% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_bridge.R
\name{updateFrameMatrixFromDF}
\alias{updateFrameMatrixFromDF}
\title{Update \code{frameMatrix} from an external source.}
\usage{
updateFrameMatrixFromDF(frameMatrix, changeDF)
}
\arguments{
\item{frameMatrix}{The \code{frameMatrix} object to be changed.}

\item{changeDF}{The data frame containing new values.}
}
\value{
An updated frameMatrix. The lower triangular matrix of \code{frameMatrix} will be read,
and the values will be flipped (e.g. 'group-inidividual' becomes 'individual-group')
and subsequently added to the upper triangular matrix.
}
\description{
Update \code{frameMatrix} from an external source.
}
\examples{
sbc007 = undupeLayers(sbc007, "trail", "name")
sbc007 = addFrameMatrix(sbc007)
#(After exporting the frame matrix and editing it)
inpath = system.file("extdata", "rez007_frame_edited.csv", package = "rezonateR")
newFrame = rez_read_csv(inpath, origDF = frameMatrix(sbc007))
frameMatrix(sbc007) = updateFrameMatrixFromDF(frameMatrix(rez007), newFrame)
}
