% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF.R
\name{new_updateFunction}
\alias{new_updateFunction}
\title{A constructor function for updateFunction objects}
\usage{
new_updateFunction(f, deps)
}
\arguments{
\item{f}{The function itself.}

\item{deps}{The dependencies of the function. Either a vector of single field names, or a single foreign field specified using an address, e.g. 'chunkDF/refexpr/word' refers to the word field of the refexpr layer of chunkDF.}
}
\value{
An updateFunction object
}
\description{
A constructor function for updateFunction objects
}
