% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_left_join.R
\name{updateLeftJoin}
\alias{updateLeftJoin}
\title{Update a field using a left join.}
\usage{
updateLeftJoin(df1, rezrObj, address, fkey, field = "")
}
\arguments{
\item{df1}{The rezrDF to be updated.}

\item{rezrObj}{The full rezrObj.}

\item{address}{An address to the field from the original DF, from the rezrObj root. For example, the 'word' field of tokenDF has the address 'tokenDF/word', and the 'word' field of the 'verb' layer of chunkDF has the address 'chunkDF/verb/word'. This may be a multiple-entry vector if you want to merge the source DFs.}

\item{fkey}{The foreign key(s). Should match the number of primary keys in the df you're pulling information from (i.e. fieldaccess set as 'key').}

\item{field}{The name of the field in the target rezrDF to be updated. If the field names in the source DFs are all the same and also the same as the name in the target DF, you may leave this unspecified.}
}
\value{
The updated data frame.
}
\description{
Update a field using a left join.
}
