% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_lowerToHigher.R
\name{updateLowerToHigher}
\alias{updateLowerToHigher}
\title{Update a field using a lowerToHigher operation.}
\usage{
updateLowerToHigher(
  df,
  rezrObj,
  address,
  fkeyAddress,
  action,
  field = "",
  fkeyInDF = FALSE,
  seqName = "docTokenSeq"
)
}
\arguments{
\item{df}{The target rezrDF to be updated.}

\item{rezrObj}{The full rezrObj.}

\item{address}{An address to the field from the original df, from the rezrObj root. For example, the 'word' field of tokenDF has the address 'tokenDF/word', and the 'word' field of the 'verb' layer of chunkDF has the address 'chunkDF/verb/word'.}

\item{fkeyAddress}{If fkeyInDF = FALSE, an address to the list of foreign keys inside the nodeMap (from the root rezrObj). If fkeyInDF = TRUE, a field in the target rezrDF containing a vector of foreign keys (not currently supported and will result in an error).}

\item{action}{The function to be performed on the values in the source rezrDF to combine them into a single value.}

\item{field}{The name of the field in the target rezrDF to be updated. If the field names in the source DFs are all the same and also the same as the name in the target DF, you may leave this unspecified.}

\item{fkeyInDF}{See fkey description.}
}
\value{
The updated data frame.
}
\description{
Not normally called by users, but acts as an updateFunct to be called by \link{reload}.
}
