% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rezrDF_easy.R, R/rezrObj_easy.R
\name{addField}
\alias{addField}
\alias{changeField}
\alias{addField.rezrDF}
\alias{changeField.rezrDF}
\alias{addField.rezrObj}
\alias{changeField.rezrObj}
\title{Shortcut functions for functions beginning with addField and changeField}
\usage{
addField(rezrObj, ..., foreign = F)

changeField(rezrObj, ..., foreign = F)

\method{addField}{rezrDF}(rezrDF, ..., foreign = F)

\method{changeField}{rezrDF}(rezrDF, ..., foreign = F)

\method{addField}{rezrObj}(rezrObj, ..., foreign = F)

\method{changeField}{rezrObj}(rezrObj, ..., foreign = F)
}
\arguments{
\item{...}{Arguments to be passed to functions such as those in see \link{addFieldLocal.rezrDF}, \link{addFieldForeign.rezrDF}, \link{addFieldLocal.rezrObj} and \link{addField.rezrObjForeign}.}

\item{foreign}{Whether you are adding a foreign field.}

\item{rezrDF}{The rezrDF to be modified.}
}
\value{
The rezrDF or rezrObj beig modified.
}
\description{
Shortcut functions for functions beginning with addField and changeField
}
\note{
This is mostly a shortcut for addLocalField and changeLocalField.
}
