% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_easy.R
\name{addFieldLocal.rezrDF}
\alias{addFieldLocal.rezrDF}
\alias{changeFieldLocal.rezrDF}
\title{Easily add a field to / change a field in a rezrDF using only information from that rezrDF}
\usage{
\method{addFieldLocal}{rezrDF}(rezrDF, fieldName, expression, fieldaccess = "flex")

\method{changeFieldLocal}{rezrDF}(rezrDF, fieldName, expression, fieldaccess = "")
}
\arguments{
\item{rezrDF}{The rezrDF object you would like to change.}

\item{fieldName}{The name of the field you would like to add or change.}

\item{expression}{An R expression. For example, if you would like to add the columns a and b of the rezrDF together, type a + b.}

\item{fieldaccess}{The access value that you would like to assign to the new or changed field.}
}
\value{
A rezrDF with the new or changed field.
}
\description{
Easily add a field to / change a field in a rezrDF using only information from that rezrDF
}
