% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_bridge.R
\name{addFrameMatrix}
\alias{addFrameMatrix}
\title{Add a frame matrix to \code{trailDF}.}
\usage{
addFrameMatrix(rezrObj)
}
\arguments{
\item{rezrObj}{The desired \code{rezrObj}}
}
\value{
A \code{rezrObj} object, but its \code{trailDF} now has an attribute which is a \code{frameMatrix} object.
}
\description{
Add a frame matrix to \code{trailDF}.
}
\note{
Make sure that the names of the trails are not duplicate with \code{undupeLayers()} first. Otherwise you will have trouble when annotating the \code{frameMatrix}.
}
\examples{
sbc007 = undupeLayers(sbc007, "trail", "name")
sbc007 = addFrameMatrix(sbc007)
}
