% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/time_seq.R
\name{addIsWordField}
\alias{addIsWordField}
\alias{addIsWordField.rezrDF}
\alias{addIsWordField.rezrObj}
\title{Add a field on whether something is a word or not.}
\usage{
addIsWordField(x, ...)

\method{addIsWordField}{rezrDF}(x, cond, addWordSeq = T)

\method{addIsWordField}{rezrObj}(x, cond, addWordSeq = T)
}
\arguments{
\item{x}{The rezrDF or rezrObj to be edited.}

\item{cond}{The wordhood condition. For example, if your word column is called 'word', and you wish to exclude zeroes, you may write 'x == "<0>"'.}

\item{addWordSeq}{If TRUE, the columns wordOrder and docWordSeq will be added.}
}
\value{
The modified \code{rezrDF} or \code{rezrObj}. If \code{addWordSeq} is set to \code{TRUE}, the columns \code{wordOrder} and \code{docWordSeq} will be added to \code{tokenDF} and \code{entryDF}, and the columns \code{wordOrderFirst}, \code{wordOrderLast}, \code{docWordSeqFirst} and \code{docWordSeqLast} will be added to \code{unitDF}, \code{chunkDF}, \code{rezDF} and \code{trackDF.}
}
\description{
Add a field on whether something is a word or not.
}
\note{
If used on a \code{rezrObj} and \code{addWordSeq = T}, \code{wordOrder} and \code{docWordSeq} are automatically added to \code{entry}, \code{unit}, \code{chunk}, \code{rez} and \code{track} tables.
}
\examples{
sbc007_withword = addIsWordField(sbc007, kind == "Word")
head(sbc007_withword$chunkDF$refexpr \%>\% select(id, text, tokenOrderFirst, wordOrderFirst, docTokenSeqLast, docWordSeqLast))
}
