% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj.R
\name{combineLayers}
\alias{combineLayers}
\alias{combineChunks}
\alias{combineTokenChunk}
\title{Combine different layers of an entity, and similar entities}
\usage{
combineLayers(rezrObj, entity, type = "intersect")

combineChunks(rezrObj, ...)

combineTokenChunk(rezrObj, type = "intersect")
}
\arguments{
\item{rezrObj}{A \code{rezrObj} object}

\item{entity}{The Rezonator entity with multiple layers (e.g. track, rez, chunk) that you want to combine}

\item{type}{The type of combination. If 'intersect', I will take the intersection of the columns in the rezrDFs. If 'union', I will take the union of the columns, populating missing fields with NAs.}

\item{...}{The rezrDF objects to be combined}
}
\value{
A \code{rezrDF} containing the required material.
}
\description{
Combine different layers of an entity, and similar entities
}
\note{
This is mainly used when you need to draw information from more than one \code{rezrDF} for the purpose of functions like \link{addFieldForeign} or \link{rez_left_join}.
}
