% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_easy.R
\name{concatenateAll}
\alias{concatenateAll}
\alias{longestLength}
\alias{longest}
\alias{shortestLength}
\alias{shortest}
\title{Functions that may be used for the complexAction field in addFieldForeign / changeFieldForeign.}
\usage{
concatenateAll(x)

longestLength(x)

longest(x)

shortestLength(x)

shortest(x)
}
\arguments{
\item{x}{The information from the source rezrDF.}
}
\value{
The aggregated data, usually a vector with a single value.
}
\description{
Functions that may be used for the complexAction field in addFieldForeign / changeFieldForeign.
}
\note{
concatenateAll concatenates everything together. longest and shortest give the longest and shortest strings, and may have multiple entries if there are ties. longestLength and shortestLength give the lengths of the longest and shortest strings in x. Some base R functions that may be used include max, min, mean, range, etc.
}
