% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_lowerToUpper.R
\name{concatStringFields}
\alias{concatStringFields}
\title{Concatenate string values in a lower-level data frame}
\usage{
concatStringFields(
  simpleDF,
  complexDF,
  complexNodeMap,
  fieldnames,
  tokenListName = "tokenList",
  ...
)
}
\arguments{
\item{simpleDF}{The lower-level dataframe, for example the token dataframe for chunks and units, or the unit dataframe for stacks.}

\item{complexDF}{The dataframe that you're trying to add the concatenated fields to.}

\item{complexNodeMap}{The node map corresponding to the simpleDF.}

\item{fieldnames}{The fields to be concatenated.}

\item{...}{Additional fields 'simpleDFAddress', 'complexNodeMapAddress', 'fieldaccess' (foreign by default) from \code{\link[=lowerToHigher]{lowerToHigher()}}. Only needed if you want automatically generated update functions.}
}
\value{
complexDF
}
\description{
Not currently available to users. I will later write a wrapper so that the user doesn't have to directly pass the DFs and node map, but can simply pass the full rez object and the names of the two DFs, and the function will also return the full rez object.
}
