% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{countCompetitors}
\alias{countCompetitors}
\title{Count the number of competing referents to the current mention}
\usage{
countCompetitors(
  cond = NULL,
  window = Inf,
  tokenOrder = NULL,
  chain = NULL,
  between = T
)
}
\arguments{
\item{cond}{The condition under which something counts as a competitor. Leave blank if anything goes.}

\item{tokenOrder}{The vector of sequence values values where the mentions appeared. Common choices are docTokenSeqFirst, docTokenSeqLast, wordTokenSeqFirst and wordTokenseqLast (the last two are available after running \link{addIsWordField} on a rezrObj. By default it's docTokenSeqLast.}

\item{chain}{The chain that each mention belongs to.}
}
\description{
This may either be counted within a window of units from the current one, or all referents competing with the current one may be counted, or a mix of both conditions. By default, we count referents intervening between the current and previous mention. Despite its name, tokenOrder can be set as unitSeqLast or similar.
}
