% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_lowerToHigher.R
\name{createLowerToHigherUpdate}
\alias{createLowerToHigherUpdate}
\title{Create an update function based on a lowerToHigher-type action.}
\usage{
createLowerToHigherUpdate(
  address,
  fkeyAddress,
  action,
  field = "",
  fkeyInDF = FALSE,
  seqName = "discourseTokenSeq"
)
}
\arguments{
\item{address}{The address to the info field that you want from your source rezrDF.}

\item{fkeyAddress}{The address to the token list in the nodeMap that corresponds to the target rezrDF.}

\item{action}{The action to be performed to combine the fields.}

\item{field}{The target field.}

\item{fkeyInDF}{Whether the foreign key will be in the data frame.}

\item{seqName}{The name of the sequencing in index in the lower rezrDF.}

\item{df}{The target rezrDF you want to modify.}

\item{rezrObj}{The rezrObj.}
}
\value{
An update function for a lowerToHigher-type foreign field.
}
\description{
A function factory that allows the user to create an update function based on a \link{lowerToHigher} operation. Not to be called by most users; \link{lowerToHigher} will call this automatically if the necessary data is supplied.
}
