% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createRezId}
\alias{createRezId}
\title{Create a Rezonator ID.}
\usage{
createRezId(n = 1, existing = character(0))
}
\arguments{
\item{n}{The number of Rezonator IDs you want.}

\item{existing}{Existing IDs, to avoid overlap.}
}
\value{
A vector of Rezonator IDs.
}
\description{
Creates a Rezonator ID when new rows need to be added. Normally does not need to be called by the user; \link{rez_add_row} and \link{addRow} do this automatically.
}
\examples{
createRezId(3)
}
