% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{getAllTreeCorrespondences}
\alias{getAllTreeCorrespondences}
\title{Relate table rows to tree entries covering the same tokens.}
\usage{
getAllTreeCorrespondences(rezrObj, entity = "chunk")
}
\arguments{
\item{rezrObj}{The rezrObj object}

\item{entity}{The name of the Rezonator entity that you will be changing - track, rez, chunk, token, unit, stack, etc.}
}
\value{
The rezrObj with table rows added to tree entries.
}
\description{
Relate table rows to tree entries covering the same tokens.
}
