% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{getRezrDFAttr}
\alias{getRezrDFAttr}
\alias{fieldaccess}
\alias{updateFunct}
\alias{inNodeMap}
\alias{setRezrDFAttr}
\alias{updateFunct<-}
\alias{fieldaccess<-}
\alias{inNodeMap<-}
\title{Extract/set an attribute of a rezrDF.}
\usage{
getRezrDFAttr(df, attr, fields = "")

fieldaccess(df, fields = "")

updateFunct(df, fields = "")

inNodeMap(df, fields = "")

setRezrDFAttr(df, attr, fields = "", value)

updateFunct(df, fields = "") <- value

fieldaccess(df, fields = "") <- value

inNodeMap(df, fields = "") <- value
}
\arguments{
\item{df}{The data.frame whose field access attributes you want to see.}

\item{attr}{The attribute that you want to extract.}

\item{fields}{The field whose access attribute you want to see. If left blank, a vector containing all the attributes is output.}

\item{value}{The value you want to set it to.}
}
\value{
The value of the attribute you want to extract (for extract functions), and the DF with the value set (for set functions).
}
\description{
Extract an attribute of a data.frame (inNodeMap, fieldaccess, updateFunct), or set it.
}
