% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_bridge.R
\name{frameMatrix}
\alias{frameMatrix}
\alias{frameMatrix<-}
\title{Extract/set a \code{frameMatrix}.}
\usage{
frameMatrix(rezrObj)

frameMatrix(rezrObj) <- value
}
\arguments{
\item{rezrObj}{}

\item{value}{The new \code{frameMatrix}.}
}
\value{
The \code{frameMatrix} of the \code{rezrObj}.
}
\description{
Extract/set a \code{frameMatrix}.
}
\examples{
frameMatrix(sbc007)[1:10, 1:12]
}
