% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate two strings together with nothing in between.}
\usage{
a \%+\% b
}
\arguments{
\item{a}{A character object.}

\item{b}{A character object.}
}
\value{
The concatenated string.
}
\description{
This is intended to be an equivalent of Python +.
}
\examples{
"3" \%+\% "a" #You get "3a"
}
