% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_tidy.R
\name{rez_group_by}
\alias{rez_group_by}
\alias{rez_ungroup}
\title{Group / ungroup rezrDFs.}
\usage{
rez_group_by(df, ...)

rez_ungroup(df, ...)
}
\arguments{
\item{df}{The data frame to be modified.}

\item{...}{Arguments to be passed to group_by / ungroup, including the field(s) on which you're grouping.}
}
\value{
The grouped / ungrouped rezrDF.
}
\description{
A replacement for \code{\link[dplyr:group_by]{dplyr::group_by()}} and \code{\link[dplyr:group_by]{dplyr::ungroup()}}.
}
\note{
There is no replacement for \code{\link[dplyr:summarise]{dplyr::summarise()}}. Currently there are no plans for a \code{rezonateR} equivalent of \code{\link[dplyr:summarise]{dplyr::summarise()}}, but please feel free to describe a use case in a GitHub issue if you would like.
}
\examples{
sbc007$tokenDF = sbc007$tokenDF \%>\%
rez_group_by(unit) \%>\% rez_mutate(unitLength = inLength(text, isWord = (kind == "Word"))) \%>\% rez_ungroup
}
