% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_importRez.R
\name{importRez}
\alias{importRez}
\title{Import a Rez file}
\usage{
importRez(paths, docnames = "", concatFields, layerRegex, separator = " ")
}
\arguments{
\item{paths}{A character vector of paths to the files to be imported. For Windows users, please use / instead of \.}

\item{docnames}{A character vector of the document names. If left blank, a docname will be generated according to the filenames of files you import. For example, the document foo/bar.rez will be named 'bar'.}

\item{concatFields}{A string of names of token-level fields, for example word or transcription, that should be concatenated to form chunk- or entry-level fields. For example, if your word field is called 'word' and you have an IPA transcription field called 'ipa', then concatFields should be c("word", "ipa").}

\item{layerRegex}{A list, each of which is a component (just track or chunk for now; stack and rez to be added later). In each list entry, there are three components: 'field' is the field on which the splitting is based; 'regex' is a vector of regular expressions; 'names' is a vector of layer names. 'regex' should have one fewer entry than 'names', as the last of the 'names' should be the default case.}

\item{separator}{The character you wish to use to separate words in concatenated columns, generally the empty string in languages like Tibetan and Chinese, and a single space in languages like Spanish and English.}
}
\value{
rezRObject
}
\description{
Import a Rez file. This returns an object containing, among other things, a nodeMap object containing raw information, and data frames for tokens, units, chunks, track chain entries, track chains, containing only key information likely to be useful for the user.
}
