% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_seq.R
\name{isInitial}
\alias{isInitial}
\alias{isFinal}
\title{Functions for getting information on whether something is located at the beginning or end of a larger structure.}
\usage{
isInitial(seq)

isFinal(seq, length)
}
\arguments{
\item{seq}{Name of the column containing the sequence value to be taken into account.}

\item{length}{Name of the column containing the maximum sequence value.}
}
\description{
Functions for getting information on whether something is located at the beginning or end of a larger structure.
}
