% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj.R
\name{getLowerFieldList}
\alias{getLowerFieldList}
\alias{getTrackTokens}
\title{Get a list of values from a lower-level table.}
\usage{
getLowerFieldList(
  rezrObj,
  fieldName,
  simpleDF,
  complexDF,
  complexNodeMap,
  listName,
  complexIDs = NULL
)

getTrackTokens(rezrObj, fieldName, trackDF = NULL)
}
\arguments{
\item{rezrObj}{The rezrObj object from which information is to be extracted.}

\item{fieldName}{The field you would like to extract.}

\item{simpleDF}{The lower rezrDF.}

\item{complexDF}{The higher rezrDF.}

\item{complexNodeMap}{The nodeMap corresponding to the higher rezrDF.}

\item{listName}{The name of the list of keys in complexNodeMap.}

\item{complexIDs}{The IDs of the rows in the complex rezrDF you would like to extract information on.}

\item{trackDF}{The trackDF you would like to extract information for.}
}
\value{
A list of vectors. Each list entry has the ID in the more complex DF as its label, and the content is a vector, each entry of which corresponds to an entry in the lower rezrDF.
}
\description{
Get a list of values from a lower-level table.
}
