% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{mergedChunksToTrack}
\alias{mergedChunksToTrack}
\title{Add merged chunks in the chunkDF to the trackDF}
\usage{
mergedChunksToTrack(rezrObj, trackLayers = NULL)
}
\arguments{
\item{rezrObj}{A rezrObj that has had chunked merged using \link{mergeChunksWithTree}.}

\item{trackLayers}{Track chain layers on which this is going to be performed. If NULL, the action will be performed on all layers.}
}
\value{
A rezrObj with the trackDF augmented with a combined chunks from the chunkDF.
}
\description{
Add merged chunks in the chunkDF to the trackDF
}
