% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRez.R
\name{new_rezrDF}
\alias{new_rezrDF}
\title{Constructor function for rezrDF}
\usage{
new_rezrDF(df, fieldaccess, updateFunct, inNodeMap)
}
\arguments{
\item{df}{The data.frame}

\item{fieldaccess}{A set of field access values}

\item{updateFunct}{A set of update functions}

\item{inNodeMap}{Whether the fields of the data.frame are in the node map}
}
\value{
A rezrDF object
}
\description{
Constructor function for rezrDF
}
