% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj.R
\name{new_rezrObj}
\alias{new_rezrObj}
\title{Constructor function for rezrObj.}
\usage{
new_rezrObj(list)
}
\arguments{
\item{list}{A node map that has been organised into nodeLists, one for each entity.}
}
\value{
The rezrObj, consisting of a series of rezrDFs and a nodeMap.
}
\description{
Not normally called by users; automatically created in \link{importRez}.
}
