% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{reload}
\alias{reload}
\alias{reload.rezrDF}
\alias{reloadLocal}
\alias{reloadForeign}
\title{Functions for reloading auto and foreign fields.}
\usage{
reload(x, ...)

\method{reload}{rezrDF}(df, rezrObj, fields = "")

reloadLocal(df, fields = "")

reloadForeign(df, rezrObj, fields = "")
}
\arguments{
\item{df}{The rezrDF you want to modify.}

\item{rezrObj}{The entire rezrObj.}

\item{fields}{The fields of the rezrDF you want to modify.}
}
\value{
The modified rezrDF. Foreign fields are always updated before auto fields. If the fields parameter is not available, I will find the best update order; otherwise, I will use the order you provide except for the rule where foreign fields are updated first.
}
\description{
Functions for reloading auto and foreign fields.
}
