% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_tidy.R
\name{rez_dfop}
\alias{rez_dfop}
\title{Perform an operation on a data.frame while updating field access status.}
\usage{
rez_dfop(df, .f, ..., fieldaccess = "flex", updateFunct = NA, oldNames = "")
}
\arguments{
\item{df}{The data frame to be modified.}

\item{.f}{The function to be performed.}

\item{...}{Argument to be passed to .f.}

\item{fieldaccess}{The field access status of the arguments.}

\item{updateFunct}{A list of updateFunctions to be added.}

\item{oldNames}{The names of the fields originally in the rezrDF AFTER the operation. If no fields will be renamed, leave blank.}
}
\value{
The DF after the operation.
}
\description{
This is a wrapper for performing data frame manipulation functions (such as dplyr functions) while also updating field access status. This function only needs to be directly called by advanced users; beginning users may stick to function such as rez_mutate and rez_left_join, which are much more simple and intuitive to use.
}
