% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_tidy.R
\name{rez_group_split}
\alias{rez_group_split}
\title{Perform a group_split on a rezrDF}
\usage{
rez_group_split(df, ...)
}
\arguments{
\item{df}{A rezrDF object.}

\item{...}{Usual group_split parameters.}
}
\value{
A list rezrDF objects after the group_split.
}
\description{
The main difference with group_split is that attributes are retained. If you use dplyr group_split, attributes will NOT be retained!
}
