% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrDF_tidy.R
\name{rez_select}
\alias{rez_select}
\title{Select columns in a rezrDF}
\usage{
rez_select(df, ...)
}
\arguments{
\item{df}{A rezrDF object}

\item{...}{Functions to be passed to dplyr select}
}
\value{
The rezrDF object with only the required columns
}
\description{
The main difference with dplyr select is that attributes will be updated to remove fields no longer there. There are no other differences at the moment.
}
\examples{
sbc007$trackDF$default \%>\% rez_select(id, name, text)
}
