% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/rezrDF_easy.R, R/rezrObj_easy.R
\name{removeField}
\alias{removeField}
\alias{removeField.rezrDF}
\alias{removeField.rezrObj}
\title{Remove a field from a \code{rezrDF}.}
\usage{
removeField(x, ...)

\method{removeField}{rezrDF}(rezrDF, fields)

\method{removeField}{rezrObj}(rezrObj, entity, layer, fields)
}
\arguments{
\item{...}{Arguments to be passed to functions such as those in see \link{addFieldLocal.rezrDF}, \link{addFieldForeign.rezrDF}, \link{addFieldLocal.rezrObj} and \link{addFieldForeign.rezrObj}.}

\item{rezrDF}{The rezrDF to be modified.}

\item{fields}{Vector of column names to be removed.}
}
\value{
A rezrDF with the fields specified in \code{fields} removed.
}
\description{
Remove a field from a \code{rezrDF}.
}
\examples{
sbc007$chunkDF$refexpr = sbc007$chunkDF$refexpr \%>\% removeField("POS_dft")
}
