% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_csv.R
\name{rez_write_csv}
\alias{rez_write_csv}
\alias{rez_read_csv}
\title{Read and write rezrDFs as CSV files}
\usage{
rez_write_csv(df, path, inclCols = character(0), exclCols = character(0), ...)

rez_read_csv(
  path,
  origDF = NULL,
  lubridate = F,
  inclCols = character(0),
  exclCols = character(0),
  ...
)
}
\arguments{
\item{df}{The rezrDF to be exported.}

\item{path}{The path from which a CSV is to be imported / exported.}

\item{inclCols}{Columns to be included from the import / export.}

\item{exclCols}{Columns to be excluded from the import / export.}

\item{...}{Arguments passed onto read_csv and write_csv.}

\item{origDF}{The rezrDF that originally produced this CSV, used to identify data types of the columns.}

\item{lubridate}{Are you using the lubridate package?}
}
\value{
For rez_read_csv, a data.frame for the CSV being imported. This is NOT a rezrDF, so please do not assign it to a rezrObj; use \link{updateFromDF} to update an existing rezrDF with the imported data.frame.
}
\description{
Replacement of readr read_csv and write_csv, but contains added functionality for dealing with rezrDFs more easily.
}
