% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezrObj.R
\name{rez_save}
\alias{rez_save}
\alias{rez_load}
\title{Save and load rezrObj and rezrDF objects.}
\usage{
rez_save(obj, filename)

rez_load(filename)
}
\arguments{
\item{obj}{The object you would like to save.}

\item{filename}{The filename, ending in \code{.Rdata}, you would like to save to.}
}
\value{
For \code{rez_load()}, the saved \code{rezrObj} or \code{rezrDF.}
}
\description{
Save and load rezrObj and rezrDF objects.
}
\examples{
rez007 = rez_load(system.file("extdata", "rez007_edit1.Rdata", package = "rezonateR"))
rez_save(rez007, "rez007_edit1.Rdata")

}
