% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{lastMentionUnit}
\alias{lastMentionUnit}
\alias{unitsToLastMention}
\alias{lastMentionToken}
\alias{tokensToLastMention}
\alias{noPrevMentions}
\alias{noPrevMentionsIf}
\alias{noPrevMentionsMatch}
\alias{getPrevMentionField}
\alias{nextMentionUnit}
\alias{unitsToNextMention}
\alias{nextMentionToken}
\alias{tokensToNextMention}
\alias{noNextMentions}
\alias{noNextMentionsIf}
\alias{noNextMentionsMatch}
\alias{getNextMentionField}
\title{Functions related to previous context in track chains.}
\usage{
lastMentionUnit(unitSeq = NULL, chain = NULL)

unitsToLastMention(unitSeq = NULL, chain = NULL)

lastMentionToken(tokenSeq = NULL, chain = NULL)

tokensToLastMention(
  tokenSeq = NULL,
  chain = NULL,
  zeroProtocol = "literal",
  zeroCond = NULL,
  unitSeq = NULL,
  unitDF = NULL
)

noPrevMentions(windowSize, unitSeq = NULL, chain = NULL)

noPrevMentionsIf(windowSize, cond, unitSeq = NULL, chain = NULL)

noPrevMentionsMatch(windowSize, field, unitSeq = NULL, chain = NULL)

getPrevMentionField(field, tokenSeq = NULL, chain = NULL)

nextMentionUnit(unitSeq = NULL, chain = NULL)

unitsToNextMention(unitSeq = NULL, chain = NULL)

nextMentionToken(tokenSeq = NULL, chain = NULL)

tokensToNextMention(
  tokenSeq = NULL,
  chain = NULL,
  zeroProtocol = "literal",
  zeroCond = NULL,
  unitSeq = NULL,
  unitDF = NULL
)

noNextMentions(windowSize, unitSeq = NULL, chain = NULL)

noNextMentionsIf(windowSize, cond, unitSeq = NULL, chain = NULL)

noNextMentionsMatch(windowSize, field, unitSeq = NULL, chain = NULL)

getNextMentionField(field, tokenSeq = NULL, chain = NULL)
}
\arguments{
\item{unitSeq}{The vector of tokenSeq values where the mentions appeared. You can choose tokenSeqFirst, tokenSeqFirst, or maybe an average of the two. By default it's tokenSeqFirst.}

\item{chain}{The chain that each mention belongs to.}

\item{cond}{For if functions, the condition that the previous / next mention must satisfy. It cannot refer to the current mention.}

\item{field}{The field whose value you want to match or extract.}
}
\description{
Functions related to previous context in track chains.
}
