% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_bridge.R
\name{updateFrameMatrixFromDF}
\alias{updateFrameMatrixFromDF}
\title{Update \code{frameMatrix} from an external source.}
\usage{
updateFrameMatrixFromDF(frameMatrix, changeDF)
}
\arguments{
\item{frameMatrix}{The \code{frameMatrix} object to be changed.}

\item{changeDF}{The data frame containing new values.}
}
\value{
An updated frameMatrix. The lower triangular matrix of \code{frameMatrix} will be read,
and the values will be flipped (e.g. 'group-inidividual' becomes 'individual-group')
and subsequently added to the upper triangular matrix.
}
\description{
Update \code{frameMatrix} from an external source.
}
